debug

+++ torchprime/torch_xla_models/train.py
@@ -481,7 +481,9 @@ class Trainer:
         segment_ids = torch.cat(
             [torch.zeros_like(segment_ids[:, :1]), segment_ids[:, :-1]], dim=1
         )
-        segment_ids = segment_ids.requires_grad_(False)
+        # Convert to float to work around scan limitation with integer tensors
+        # See: https://github.com/pytorch/xla/issues/8783
+        segment_ids = segment_ids.float().requires_grad_(False)
         batch["segment_ids"] = segment_ids
         logger.info(f"input_ids: {batch['input_ids']}")
         logger.info(f"segment_ids: {batch['segment_ids']}")
